+++ src/Elders.Cronus.Projections.Cassandra/CasssandraCollectionPersister.cs
@@ -1,43 +1,11 @@
-﻿using System;
-using System.Collections.Concurrent;
+﻿using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using Cassandra;
-using Elders.Cronus.DomainModeling;
 using Elders.Cronus.DomainModeling.Projections;
 
 namespace Elders.Cronus.Projections.Cassandra
 {
-    public static class CasssandraCollectionPersisterExtensions
-    {
-        private const string CreateKeyValueDataColumnFamilyTemplate = @"CREATE TABLE IF NOT EXISTS ""{0}"" (id text, data blob, PRIMARY KEY (id)) WITH compression = {{ 'sstable_compression' : '' }};";
-        private const string CreateColumnFamilyTemplate = @"CREATE TABLE IF NOT EXISTS ""{0}"" (id text, iid text, data blob, PRIMARY KEY (id,iid)) WITH compression = {{ 'sstable_compression' : '' }};";
-
-        public static void InitializeProjectionDatabase(this ISession session, IEnumerable<Type> projectionTypes)
-        {
-            foreach (var projType in projectionTypes.Where(x => x.GetInterfaces().Any(y => y.IsGenericType && y.GetGenericTypeDefinition() == typeof(IDataTransferObject<>))))
-            {
-                session.Execute(string.Format(CreateKeyValueDataColumnFamilyTemplate, projType.GetColumnFamily()).ToLowerInvariant());
-            }
-
-            foreach (var projType in projectionTypes.Where(x => x.GetInterfaces().Any(y => y.IsGenericType && y.GetGenericTypeDefinition() == typeof(ICollectionDataTransferObjectItem<,>))))
-            {
-                session.Execute(string.Format(CreateColumnFamilyTemplate, projType.GetColumnFamily()).ToLowerInvariant());
-            }
-        }
-
-        public static void InitializeProjectionDatabase(this ISession session, Assembly assemblyContainingProjections)
-        {
-            InitializeProjectionDatabase(session, assemblyContainingProjections.GetExportedTypes());
-        }
-
-        public static string GetColumnFamily(this Type projectionType)
-        {
-            return projectionType.GetContractId().Replace("-", "").ToLowerInvariant();
-        }
-    }
-
     public class CasssandraCollectionPersister : IKeyValueCollectionPersister
     {
         private const string InsertQueryTemplate = @"INSERT INTO ""{0}"" (id,iid,data) VALUES (?,?,?);";
@@ -60,11 +28,17 @@ namespace Elders.Cronus.Projections.Cassandra
 
         private readonly ConcurrentDictionary<string, PreparedStatement> DeletePreparedStatements;
 
+        private readonly ConsistencyLevel writeConsistencyLevel;
+
+        private readonly ConsistencyLevel readConsistencyLevel;
+
         private ISession session;
 
-        public CasssandraCollectionPersister(ISession session)
+        public CasssandraCollectionPersister(ISession session, ConsistencyLevel writeConsistencyLevel, ConsistencyLevel readConsistencyLevel)
         {
             this.session = session;
+            this.writeConsistencyLevel = writeConsistencyLevel;
+            this.readConsistencyLevel = readConsistencyLevel;
             this.SavePreparedStatements = new ConcurrentDictionary<string, PreparedStatement>();
             this.GetPreparedStatements = new ConcurrentDictionary<string, PreparedStatement>();
             this.GetItemPreparedStatements = new ConcurrentDictionary<string, PreparedStatement>();
@@ -75,6 +49,7 @@ namespace Elders.Cronus.Projections.Cassandra
         public IEnumerable<KeyValueCollectionItem> GetCollection(string collectionId, string columnFamily)
         {
             var statement = GetPreparedStatements.GetOrAdd(columnFamily, x => BuildeGetPreparedStatemnt(x));
+            statement.SetConsistencyLevel(readConsistencyLevel);
             var result = session.Execute(statement.Bind(collectionId));
             foreach (var row in result)
             {
@@ -85,6 +60,7 @@ namespace Elders.Cronus.Projections.Cassandra
         public KeyValueCollectionItem GetCollectionItem(string collectionId, string itemId, string columnFamily)
         {
             var statement = GetItemPreparedStatements.GetOrAdd(columnFamily, x => BuildeGetCollectionItemPreparedStatemnt(x));
+            statement.SetConsistencyLevel(readConsistencyLevel);
             var result = session.Execute(statement.Bind(collectionId, itemId)).FirstOrDefault();
             if (result == null)
                 return null;
@@ -95,12 +71,21 @@ namespace Elders.Cronus.Projections.Cassandra
         public void AddToCollection(KeyValueCollectionItem collectionItem)
         {
             var statement = SavePreparedStatements.GetOrAdd(collectionItem.Table, x => BuildeInsertPreparedStatemnt(x));
+            statement.SetConsistencyLevel(writeConsistencyLevel);
             var result = session.Execute(statement.Bind(collectionItem.CollectionId, collectionItem.ItemId, collectionItem.Blob));
+        }
+
+        public void Update(KeyValueCollectionItem collectionItem, byte[] data)
+        {
+            var statement = UpdatePreparedStatements.GetOrAdd(collectionItem.Table, x => BuildeInsertPreparedStatemnt(x));
+            statement.SetConsistencyLevel(writeConsistencyLevel);
+            var result = session.Execute(statement.Bind(collectionItem.CollectionId, collectionItem.ItemId, data));
         }
 
         public void DeleteCollectionItem(KeyValueCollectionItem collectionItem)
         {
             var statement = DeletePreparedStatements.GetOrAdd(collectionItem.Table, x => BuildeDeletePreparedStatemnt(x));
+            statement.SetConsistencyLevel(writeConsistencyLevel);
             var result = session.Execute(statement.Bind(collectionItem.CollectionId, collectionItem.ItemId));
         }
 
@@ -122,12 +107,6 @@ namespace Elders.Cronus.Projections.Cassandra
         private PreparedStatement BuildeDeletePreparedStatemnt(string columnFamily)
         {
             return session.Prepare(string.Format(DeleteQueryTemplate, columnFamily));
-        }
-
-        public void Update(KeyValueCollectionItem collectionItem, byte[] data)
-        {
-            var statement = UpdatePreparedStatements.GetOrAdd(collectionItem.Table, x => BuildeInsertPreparedStatemnt(x));
-            var result = session.Execute(statement.Bind(collectionItem.CollectionId, collectionItem.ItemId, data));
         }
 
         private PreparedStatement BuildeUpdatePreparedStatemnt(string columnFamily)