+++ src/Elders.Cronus.Projections.Cassandra/EventSourcing/IProjectionRepository.cs
@@ -1,0 +1,9 @@
+﻿using Elders.Cronus.DomainModeling;
+
+namespace Elders.Cronus.Projections.Cassandra.EventSourcing
+{
+    public interface IProjectionRepository
+    {
+        IProjectionGetResult<T> Get<T>(IBlobId projectionId) where T : IProjectionDefinition;
+    }
+}